feat: Migrate to Route Groups for Improved Navigation and Performance
- Organized routes into logical groups to enhance project structure without affecting URL paths.
- Implemented Route Groups to allow for nested layouts at the same segment level, enabling multiple layouts within a common segment.
- Improved the Sidebar Dashboard functionality to maintain state across route changes, eliminating the need for reloads on every navigation.
- Enhanced loading performance by preparing for server-side rendering in Next.js, allowing the Sidebar to load independently while pages automatically handle loading states.
