+++ frontend/package.json
@@ -14,6 +14,7 @@
     "chart.js": "^4.4.5",
     "clsx": "^2.1.1",
     "framer-motion": "^11.11.9",
+    "mini-svg-data-uri": "^1.4.4",
     "next": "14.2.16",
     "react": "^18",
     "react-chartjs-2": "^5.2.0",
