+++ frontend/src/app/(user-dashboard)/user-dashboard/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+import { BentoGridDemo } from "@/components/dashboard";
+import React from "react";
+
+function Dashboard() {
+  return (
+    <div className="flex-1">
+      <BentoGridDemo />
+    </div>
+  );
+}
+
+export default Dashboard;
