+++ package.json
@@ -1,0 +1,5 @@
+{
+	"dependencies": {
+		"mini-svg-data-uri": "^1.4.4"
+	}
+}
